Moved email sending file to utility
